Clean up ObservableCollectionEvent signal extensions and add diff operator.
